tools/requirements.txt
# Python Requirements for trace-scope Tools
#
# All tools consolidated into TWO files:
# - trc.py: All-in-one tool (instrumentation, analysis, stats, callgraph, compare, diff)
# - test_trc.py: Unified test suite (53 tests)
#
# Uses ONLY Python standard library - no external dependencies required!
#
# Standard library modules used:
# - argparse (command-line parsing)
# - struct (binary format reading/writing)
# - re (regular expressions for wildcard matching)
# - sys, os (system utilities)
# - tempfile (testing)
# - csv, json (export formats)
# - collections (defaultdict for call graphs)
# - unittest (testing framework)
#
# Minimum Python version: 3.6+
#
# Usage:
#   python trc.py <command> [options]     # Run tool
#   python test_trc.py                    # Run tests
#
# Optional (for enhanced call graph visualization):
# graphviz>=0.20  # For rendering DOT files to PNG/SVG/PDF
#                 # Install: pip install graphviz
#                 # Note: Also requires GraphViz system package
#                 #       Windows: choco install graphviz
#                 #       Linux: apt-get install graphviz
#                 #       macOS: brew install graphviz
#
# The tools work perfectly without graphviz - you just need to manually
# render DOT files using the 'dot' command-line tool:
#   dot -Tpng callgraph.dot -o callgraph.png

# Required for modular header extraction and merging:
libclang>=18.0  # For C++ AST parsing and validation

